# Fix panic when connection dropped in the middle of the sid