Fixes #418. Includes a unit test that exhibits the panic.

Q: is errors.Is(err, ErrSSHFxConnectionLost) the right check in the test?