@puellanivis 2021-03-16
For a current look at what’s driving me absolutely bonkers:

```go
package narrow

import (
	"fmt"
)

type S struct{
	id uint32
}

func (s S) ID() uint32 { return s.id }

type ider interface{
	ID() uint32
}

func Func(p ider) {
	fmt.Println(p.ID())
}

func Harness() {
	Func(S{
		id: 42,
	})
}
```

```
$ ( go test -gcflags="-m" 2>&1)
…
./narrow.go:22:8: S literal escapes to heap
…
```