@puellanivis 2021-03-16
So, I’ve gotten things to a “stable state”. The benchmarks have actually changed so much that they’re no longer “valid” in the old version. I might just push a PR for that alone, so we can get a clean comparison.

Basically, I used the `-memprofile` flag to dump memory stats, and shaved down actual bottlenecks. A lot of the benchmarks were actually doing a lot of work and memory allocation outside of the actual execution. The newer version will have cleaner benchmarks that aren’t doing extra work in the benchmark. The benchmarks are laser focused on what they’re actually trying to measure.

After that it’s cleaning up the code I have. I’m pretty sure it’s now at like 5 different layers of redundant code in some places. 😆 So, this one is gonna be a big refactor. Basically, pulling out most of the unmarshaling guts and replacing it. 🤯