@greatroar 2021-03-18
@puellanivis I wonder how far you got with the unmarshaling? I took a stab at it myself, and while I don't have a branch up yet (too many failing tests), I do have a design of an unmarshaler for the client that you might find interesting. It has a new type,

```go
type receivedPacket struct {
	bytes.Buffer
	Type uint8
	// Two more fields for optimization/sid handling.
}

func recvPacket(r io.Reader) (*receivedPacket, error) {
	/* Get a receivedPacket from recvPool, fills it from r. */
}

var recvPool = sync.Pool{
	New: func() interface{} { return &receivedPacket{} },
}

func (p *receivedPacket) release() {
	p.Reset()
	recvPool.Put(p)
}

// Unmarshal methods advance the read offset on p.Buffer.
// One sftp.Client method uses binary.Read, which works on *receivedPacket directly.

func (p *receivedPacket) unmarshalExtensionPair() (ep extensionPair, err error) {
	ep.Name, err = p.unmarshalString()
	ep.Data, err = p.unmarshalString()
	return ep, err
}
func (p *receivedPacket) unmarshalString() (string, error) {
	l, _ := p.unmarshalUint32()
	b := p.Next(int(l))
	if len(b) < int(l) {
		return "", errShortPacket
	}
	return string(b), nil
}

func (p *receivedPacket) unmarshalUint32() (uint32, error) {
	b := p.Next(4)
	if len(b) < 4 {
		return 0, errShortPacket
	}
	return binary.BigEndian.Uint32(b), nil
}
```

Plugging this into sftp.Client methods was pretty easy. For the servers, the UnmarshalBinary methods should be replaceable by methods that read from a receivedPacket. Their error handling is much less in the way, since the receivedPacket.unmarshal* will just keep returning errShortPacket once one of them has hit the end; see unmarshalExtensionPair, above.