@greatroar 2021-03-18
Not everything works yet, but I am getting some nice preliminary benchmark results:

```
name                       old time/op    new time/op    delta
WriteTo1k-8                  12.7ms ± 2%     9.9ms ± 3%  -21.87%  (p=0.007 n=10+3)
WriteTo16k-8                 12.6ms ± 3%     9.9ms ± 3%  -21.08%  (p=0.007 n=10+3)
WriteTo32k-8                 12.6ms ± 3%     9.9ms ± 1%  -21.37%  (p=0.007 n=10+3)
WriteTo128k-8                12.5ms ± 3%     9.8ms ± 1%  -21.32%  (p=0.007 n=10+3)
WriteTo512k-8                12.6ms ± 2%     9.9ms ± 1%  -21.20%  (p=0.007 n=10+3)
WriteTo1MiB-8                12.6ms ± 3%    10.0ms ± 5%  -20.65%  (p=0.007 n=10+3)
WriteTo4MiB-8                12.6ms ± 4%    10.0ms ± 4%  -20.60%  (p=0.007 n=10+3)
WriteTo4MiBDelay10Msec-8      109ms ± 1%     107ms ± 0%   -1.25%  (p=0.007 n=10+3)
WriteTo4MiBDelay50Msec-8      508ms ± 0%     508ms ± 0%     ~     (p=1.000 n=9+3)
WriteTo4MiBDelay150Msec-8     1.51s ± 0%     1.51s ± 0%     ~     (p=0.209 n=9+3)
Read1k-8                      225ms ± 4%     178ms ±13%  -21.16%  (p=0.001 n=10+5)
Read16k-8                    26.2ms ± 5%    20.8ms ± 9%  -20.47%  (p=0.007 n=10+3)
Read32k-8                    19.1ms ± 2%    15.2ms ± 1%  -20.29%  (p=0.007 n=10+3)
Read128k-8                   10.9ms ± 2%     7.5ms ± 1%  -30.64%  (p=0.007 n=10+3)
Read512k-8                   8.36ms ± 2%    4.80ms ± 1%  -42.65%  (p=0.007 n=10+3)
Read1MiB-8                   8.57ms ± 2%    4.75ms ± 1%  -44.53%  (p=0.007 n=10+3)
Read4MiB-8                   9.80ms ± 2%    5.44ms ± 1%  -44.48%  (p=0.007 n=10+3)
Read4MiBDelay10Msec-8        72.3ms ± 1%    68.9ms ± 1%   -4.65%  (p=0.007 n=10+3)
Read4MiBDelay50Msec-8         311ms ± 0%     309ms ± 0%   -0.83%  (p=0.007 n=10+3)
Read4MiBDelay150Msec-8        912ms ± 0%     910ms ± 0%   -0.22%  (p=0.024 n=6+3)

name                       old speed      new speed      delta
WriteTo1k-8                 828MB/s ± 2%  1060MB/s ± 3%  +28.02%  (p=0.007 n=10+3)
WriteTo16k-8                833MB/s ± 3%  1056MB/s ± 3%  +26.77%  (p=0.007 n=10+3)
WriteTo32k-8                828MB/s ± 2%  1057MB/s ± 1%  +27.65%  (p=0.009 n=9+3)
WriteTo128k-8               839MB/s ± 3%  1067MB/s ± 1%  +27.06%  (p=0.007 n=10+3)
WriteTo512k-8               831MB/s ± 2%  1055MB/s ± 1%  +26.89%  (p=0.007 n=10+3)
WriteTo1MiB-8               832MB/s ± 3%  1049MB/s ± 5%  +26.14%  (p=0.007 n=10+3)
WriteTo4MiB-8               830MB/s ± 4%  1047MB/s ± 4%  +26.01%  (p=0.007 n=10+3)
WriteTo4MiBDelay10Msec-8   96.5MB/s ± 1%  97.8MB/s ± 0%   +1.26%  (p=0.007 n=10+3)
WriteTo4MiBDelay50Msec-8   20.6MB/s ± 0%  20.6MB/s ± 0%     ~     (p=0.864 n=9+3)
WriteTo4MiBDelay150Msec-8  6.94MB/s ± 0%  6.94MB/s ± 0%   -0.05%  (p=0.000 n=7+3)
Read1k-8                   46.5MB/s ± 4%  59.5MB/s ±12%  +27.88%  (p=0.001 n=10+5)
Read16k-8                   401MB/s ± 5%   506MB/s ± 9%  +26.18%  (p=0.007 n=10+3)
Read32k-8                   549MB/s ± 2%   689MB/s ± 1%  +25.43%  (p=0.007 n=10+3)
Read128k-8                  964MB/s ± 2%  1390MB/s ± 1%  +44.16%  (p=0.007 n=10+3)
Read512k-8                 1.25GB/s ± 2%  2.19GB/s ± 1%  +74.36%  (p=0.007 n=10+3)
Read1MiB-8                 1.22GB/s ± 2%  2.21GB/s ± 1%  +80.27%  (p=0.007 n=10+3)
Read4MiB-8                 1.07GB/s ± 2%  1.93GB/s ± 1%  +80.11%  (p=0.007 n=10+3)
Read4MiBDelay10Msec-8       145MB/s ± 1%   152MB/s ± 1%   +4.88%  (p=0.007 n=10+3)
Read4MiBDelay50Msec-8      33.7MB/s ± 0%  34.0MB/s ± 0%   +0.84%  (p=0.007 n=10+3)
Read4MiBDelay150Msec-8     11.5MB/s ± 0%  11.5MB/s ± 0%   +0.22%  (p=0.007 n=10+3)

name                       old alloc/op   new alloc/op   delta
WriteTo1k-8                  15.8MB ± 0%     3.3MB ± 0%  -79.04%  (p=0.007 n=10+3)
WriteTo16k-8                 15.9MB ± 0%     3.4MB ± 1%  -78.83%  (p=0.007 n=10+3)
WriteTo32k-8                 15.8MB ± 0%     3.3MB ± 0%  -78.86%  (p=0.007 n=10+3)
WriteTo128k-8                15.8MB ± 0%     3.3MB ± 0%  -78.89%  (p=0.007 n=10+3)
WriteTo512k-8                15.8MB ± 0%     3.4MB ± 1%  -78.86%  (p=0.012 n=8+3)
WriteTo1MiB-8                15.8MB ± 0%     3.3MB ± 1%  -78.91%  (p=0.007 n=10+3)
WriteTo4MiB-8                15.9MB ± 0%     3.4MB ± 3%  -78.71%  (p=0.009 n=9+3)
WriteTo4MiBDelay10Msec-8     18.8MB ± 0%     6.7MB ± 0%  -64.29%  (p=0.007 n=10+3)
WriteTo4MiBDelay50Msec-8     32.1MB ± 0%    20.6MB ± 0%  -35.98%  (p=0.007 n=10+3)
WriteTo4MiBDelay150Msec-8    48.8MB ± 0%    37.7MB ± 2%  -22.71%  (p=0.017 n=7+3)
Read1k-8                     6.99MB ± 0%    0.95MB ± 0%  -86.41%  (p=0.007 n=10+3)
Read16k-8                    5.99MB ± 0%    0.06MB ± 1%  -98.92%  (p=0.000 n=10+3)
Read32k-8                    6.63MB ± 0%    0.04MB ± 8%  -99.45%  (p=0.007 n=10+3)
Read128k-8                   6.77MB ± 0%    0.06MB ± 8%  -99.10%  (p=0.009 n=9+3)
Read512k-8                   7.25MB ± 0%    0.06MB ±11%  -99.23%  (p=0.007 n=10+3)
Read1MiB-8                   7.91MB ± 0%    0.06MB ± 4%  -99.29%  (p=0.007 n=10+3)
Read4MiB-8                   10.5MB ± 0%     0.1MB ±24%  -99.50%  (p=0.012 n=8+3)
Read4MiBDelay10Msec-8        10.5MB ± 0%     0.1MB ± 8%  -98.79%  (p=0.007 n=10+3)
Read4MiBDelay50Msec-8        10.5MB ± 0%     0.4MB ± 0%  -96.52%  (p=0.007 n=10+3)
Read4MiBDelay150Msec-8       10.5MB ± 0%     0.6MB ±13%  -93.84%  (p=0.007 n=10+3)

name                       old allocs/op  new allocs/op  delta
WriteTo1k-8                   2.71k ± 0%     1.67k ± 0%  -38.40%  (p=0.000 n=10+3)
WriteTo16k-8                  2.71k ± 0%     1.67k ± 0%  -38.27%  (p=0.000 n=10+3)
WriteTo32k-8                  2.71k ± 0%     1.67k ± 0%  -38.35%  (p=0.000 n=10+3)
WriteTo128k-8                 2.71k ± 0%     1.67k ± 0%  -38.41%  (p=0.000 n=10+3)
WriteTo512k-8                 2.71k ± 0%     1.67k ± 0%  -38.36%  (p=0.000 n=10+3)
WriteTo1MiB-8                 2.71k ± 0%     1.67k ± 0%  -38.35%  (p=0.007 n=10+3)
WriteTo4MiB-8                 2.71k ± 0%     1.67k ± 0%  -38.30%  (p=0.000 n=9+3)
WriteTo4MiBDelay10Msec-8      3.11k ± 0%     2.08k ± 0%  -33.01%  (p=0.009 n=9+3)
WriteTo4MiBDelay50Msec-8      3.13k ± 1%     2.16k ± 1%  -31.13%  (p=0.009 n=9+3)
WriteTo4MiBDelay150Msec-8     3.14k ± 0%     2.22k ± 2%  -29.49%  (p=0.012 n=8+3)
Read1k-8                      30.7k ± 0%     20.5k ± 0%  -33.31%  (p=0.000 n=9+3)
Read16k-8                     1.94k ± 0%     1.29k ± 0%  -33.32%  (p=0.000 n=10+3)
Read32k-8                       982 ± 0%       655 ± 0%  -33.30%  (p=0.000 n=10+3)
Read128k-8                    1.25k ± 0%     0.91k ± 0%  -26.72%  (p=0.000 n=9+3)
Read512k-8                    1.14k ± 0%     0.78k ± 0%  -31.45%  (p=0.007 n=10+3)
Read1MiB-8                    1.22k ± 0%     0.82k ± 0%  -32.83%  (p=0.000 n=10+3)
Read4MiB-8                    1.31k ± 0%     0.79k ± 0%  -39.75%  (p=0.000 n=10+3)
Read4MiBDelay10Msec-8         1.57k ± 0%     1.05k ± 0%  -32.89%  (p=0.007 n=10+3)
Read4MiBDelay50Msec-8         1.57k ± 1%     1.09k ± 1%  -30.89%  (p=0.007 n=10+3)
Read4MiBDelay150Msec-8        1.58k ± 1%     1.12k ± 1%  -29.30%  (p=0.007 n=10+3)
```